add job and controller for collecting data via http get requests

+++ wevi/database/migrations/2016_09_24_105729_create_datapoints_table.php
@@ -18,7 +18,7 @@ class CreateDatapointsTable extends Migration
             $table->string('station');
             $table->string('key');
             $table->double('value');
-            $table->integer('data_timestamp');
+            $table->double('data_timestamp');
             $table->timestamps();
         });
     }
